[Modified]指摘箇所を修正
1. ポイントの変数名をscoreに統一
2. スペアメソッドを追加
3. メソッドを切り出したファイルを削除
4. スペルミスを修正 (x: culc => o: calc)
